+++ utils/consts.ts
@@ -1,0 +1,17 @@
+const VALID_CAMPAIGNS = [{ value: 1, label: 'betting-integrations' }]
+
+const VALID_TAGS = [
+  { value: 1, label: 'Tag 1' },
+  { value: 2, label: 'Tag 2' },
+  { value: 3, label: 'Tag 3' },
+]
+
+const VALID_CHANNELS = [
+  { value: 1, label: 'espn-content' },
+  { value: 2, label: 'espn-app' },
+  { value: 3, label: 'espn-web' },
+  { value: 4, label: 'espn-fantasy-app' },
+  { value: 5, label: 'espn-fantasy-web' },
+]
+
+export { VALID_CAMPAIGNS, VALID_TAGS, VALID_CHANNELS }
